Move BlivetResetTestCases from ImageBackedTestCase to VMBackedTestCase
